# apps/backend/property-service/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
			"LocalizedContent": {
				"properties": {
					"content": {
						"type": "string"
					},
					"language": {
						"type": "string"
					}
				},
				"required": [
					"content",
					"language"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"ResponseCreatePropertyDTO": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"cognitoSub": {
						"type": "string"
					},
					"title": {
						"items": {
							"$ref": "#/components/schemas/LocalizedContent"
						},
						"type": "array"
					},
					"description": {
						"items": {
							"$ref": "#/components/schemas/LocalizedContent"
						},
						"type": "array"
					},
					"thumbnail": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"urlmap": {
						"type": "string"
					},
					"address": {
						"items": {
							"$ref": "#/components/schemas/LocalizedContent"
						},
						"type": "array"
					},
					"location": {
						"items": {
							"$ref": "#/components/schemas/LocalizedContent"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"items": {
							"$ref": "#/components/schemas/LocalizedContent"
						},
						"type": "array"
					},
					"transition": {
						"items": {
							"$ref": "#/components/schemas/LocalizedContent"
						},
						"type": "array"
					},
					"detail": {
						"$ref": "#/components/schemas/Record_string.any_"
					},
					"status": {
						"type": "boolean"
					},
					"condanate": {
						"items": {
							"$ref": "#/components/schemas/IProperty"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"cognitoSub",
					"thumbnail",
					"images",
					"location",
					"category",
					"transition"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pagination": {
				"properties": {
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"totalProperty": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"currentPage",
					"totalPages",
					"totalProperty"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseAllPropertyDTO": {
				"properties": {
					"properties": {
						"items": {
							"$ref": "#/components/schemas/ResponseCreatePropertyDTO"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/Pagination"
					}
				},
				"required": [
					"properties",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "property-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/api/v1/properties": {
			"post": {
				"operationId": "CreateProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseCreatePropertyDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"thumbnail": {
										"type": "string",
										"format": "binary"
									},
									"images": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									},
									"title": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"urlmap": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"location": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"transition": {
										"type": "string"
									},
									"detail": {
										"type": "string"
									},
									"coordinate": {
										"type": "string"
									}
								},
								"required": [
									"thumbnail",
									"images",
									"title",
									"description"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseAllPropertyDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cognitoSub",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "title",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "description",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "address",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transition",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "price",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "price_gte",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "price_lte",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 12,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/properties/get/{propertyId}": {
			"get": {
				"operationId": "FetchPropertyByID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePropertyDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/properties/me": {
			"get": {
				"operationId": "GetPropertyMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseAllPropertyDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "title",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "description",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "address",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transition",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "price",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "price_gte",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "price_lte",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 12,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "fav_me",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/properties/me/{propertyId}": {
			"put": {
				"operationId": "UpdateProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseUpdatePropertyDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"thumbnail": {
										"type": "string",
										"format": "binary"
									},
									"images": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									},
									"title": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"urlmap": {
										"type": "string"
									},
									"address": {
										"type": "string"
									},
									"location": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"transition": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"detail": {
										"type": "string"
									},
									"status": {
										"type": "string"
									},
									"coordinate": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/properties/{propertyId}/views": {
			"put": {
				"operationId": "IncrementPropertyViews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePropertyDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetPropertyViews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"views": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"views"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/properties/{propertyId}/status": {
			"put": {
				"operationId": "UpdatePropertyStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedStatus": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"type": "boolean"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/properties/{propertyId}/statusAdmin": {
			"put": {
				"operationId": "UpdateAdminstatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatestatusAdmin": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"statusAdmin": {
										"type": "boolean"
									}
								},
								"required": [
									"statusAdmin"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/properties/{propertyId}": {
			"delete": {
				"operationId": "DeletePropertyById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/properties/user/{cognitoSub}": {
			"get": {
				"operationId": "GetPropertyUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalProperties": {
											"type": "number",
											"format": "double"
										},
										"totalPages": {
											"type": "number",
											"format": "double"
										},
										"properties": {
											"items": {
												"$ref": "#/components/schemas/ResponsePropertyDTO"
											},
											"type": "array"
										}
									},
									"required": [
										"totalProperties",
										"totalPages",
										"properties"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cognitoSub",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cognitoSub",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "title",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "description",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "address",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transition",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "price",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "price_gte",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "price_lte",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/properties/{propertyId}/coordinates": {
			"post": {
				"operationId": "AddCoordinatesToProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePropertyDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"lng": {
										"type": "number",
										"format": "double"
									},
									"lat": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"lng",
									"lat"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/properties/nearby": {
			"get": {
				"operationId": "FindNearbyProperties",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ResponsePropertyDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Property Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "lat",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "lng",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxDistance",
						"required": false,
						"schema": {
							"default": 1000,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}